% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Knights_Tour.R
\name{kt_valid_move}
\alias{kt_valid_move}
\title{Does the next position help make a knight's tour?}
\usage{
kt_valid_move(position = position, solution)
}
\arguments{
\item{position}{A numeric vector of length 2 describing the coordinates of
the proposed next position on a chess board.}

\item{solution}{A numeric 8x8 matrix describing the current partial solution
on a chess board.}
}
\value{
\code{TRUE} if the proposed next position can be appended to the
  given partial knight's tour to form another knight's tour (which may or may
  not be a complete knight's tour). Otherwise \code{FALSE}.
}
\description{
Can the proposed next position be appended to the given partial knight's tour
  to create a larger knight's tour (which may or may not be a complete
  knight's tour)?
}
